scantest: fixes for new API

+++ tests/scantest.py
@@ -1,14 +1,14 @@
 #!/usr/bin/python
 
-import numpy, os, time, cv, sys, math, sys
+import numpy, os, time, cv, sys, math, sys, glob
 
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.realpath(__file__)), '..', 'image'))
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.realpath(__file__)), '..', 'camera'))
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.realpath(__file__)), '..', 'lib'))
-import scanner, cuav_util, cuav_mosaic, mav_position, chameleon, cuav_joe, mp_slipmap
+import scanner, cuav_util, cuav_mosaic, mav_position, chameleon, cuav_joe, mp_slipmap, mp_image
 
 from optparse import OptionParser
-parser = OptionParser("scantest.py [options] <filename..>")
+parser = OptionParser("scantest.py [options] <directory>")
 parser.add_option("--repeat", type='int', default=1, help="scan repeat count")
 parser.add_option("--view", action='store_true', default=False, help="show images")
 parser.add_option("--fullres", action='store_true', default=False, help="show full resolution")
@@ -19,29 +19,41 @@ parser.add_option("--quality", type='int', default=95, help="jpeg compression qu
 parser.add_option("--mosaic", action='store_true', default=False, help="build a mosaic of regions")
 parser.add_option("--mavlog", default=None, help="flight log for geo-referencing")
 parser.add_option("--boundary", default=None, help="search boundary file")
-parser.add_option("--max-deltat", default=1.0, type='float', help="max deltat for interpolation")
+parser.add_option("--max-deltat", default=0.0, type='float', help="max deltat for interpolation")
 parser.add_option("--max-attitude", default=45, type='float', help="max attitude geo-referencing")
 parser.add_option("--joe", default=None, help="file containing list of joe positions")
 parser.add_option("--linkjoe", default=None, help="link joe images to this directory")
 parser.add_option("--show-misses", default=False, action='store_true', help="show missed Joes")
 parser.add_option("--lens", default=4.0, type='float', help="lens focal length")
+parser.add_option("--roll-stabilised", default=False, action='store_true', help="roll is stabilised")
+parser.add_option("--gps-lag", default=0.0, type='float', help="GPS lag in seconds")
 (opts, args) = parser.parse_args()
 
 class state():
   def __init__(self):
     pass
 
+slipmap = None
 
-def process(files):
+def process(args):
   '''process a set of files'''
 
+  global slipmap
   scan_count = 0
+  files = []
+  for a in args:
+    if os.path.isdir(a):
+      files.extend(glob.glob(os.path.join(a, '*.pgm')))
+    else:
+      files.append(a)
+  files.sort()
   num_files = len(files)
+  print("num_files=%u" % num_files)
   region_count = 0
   joes = []
 
   if opts.mavlog:
-    mpos = mav_position.MavInterpolator()
+    mpos = mav_position.MavInterpolator(gps_lag=opts.gps_lag)
     mpos.set_logfile(opts.mavlog)
   else:
     mpos = None
@@ -53,29 +65,51 @@ def process(files):
 
   if opts.mosaic:
     slipmap = mp_slipmap.MPSlipMap(service='GoogleSat', elevation=True, title='Map')
+    icon = slipmap.icon('planetracker.png')
+    slipmap.add_object(mp_slipmap.SlipIcon('plane', (0,0), icon, layer=3, rotation=0,
+                                           follow=True,
+                                           trail=mp_slipmap.SlipTrail()))
     mosaic = cuav_mosaic.Mosaic(slipmap, lens=opts.lens)
     if boundary is not None:
       mosaic.set_boundary(boundary)
 
   if opts.joe:
-    joes = cuav_util.load_joes(opts.joe)
+    joes = cuav_util.polygon_load(opts.joe)
     if boundary:
-      for joe in joes:
+      for i in range(len(joes)):
+        joe = joes[i]
         if cuav_util.polygon_outside(joe, boundary):
           print("Error: joe outside boundary", joe)
           return
+        icon = slipmap.icon('flag.png')
+        slipmap.add_object(mp_slipmap.SlipIcon('joe%u' % i, (joe[0],joe[1]), icon, layer=4))
 
   joelog = cuav_joe.JoeLog('joe.log')      
 
+  if opts.view:
+    viewer = mp_image.MPImage(title='Image')
+
   for f in files:
+    time.sleep(0.3)
     frame_time = cuav_util.parse_frame_time(f)
     if mpos:
       try:
-        pos = mpos.position(frame_time, opts.max_deltat)
-      except mav_position.MavInterpolatorException:
+        if opts.roll_stabilised:
+          roll = 0
+        else:
+          roll = None
+        pos = mpos.position(frame_time, opts.max_deltat,roll=roll)
+        slipmap.set_position('plane', (pos.lat, pos.lon), rotation=pos.yaw)
+      except mav_position.MavInterpolatorException as e:
+        print e
         pos = None
     else:
       pos = None
+
+    # check for any events from the map
+    if opts.mosaic:
+      slipmap.check_events()
+
     if f.endswith('.pgm'):
       pgm = cuav_util.PGM(f)
       im = pgm.array
@@ -118,19 +152,22 @@ def process(files):
 
     # optionally link all the images with joe into a separate directory
     # for faster re-running of the test with just joe images
-    if opts.linkjoe and len(regions) > 0:
+    if pos and opts.linkjoe and len(regions) > 0:
       cuav_util.mkdir_p(opts.linkjoe)
-      joepath = os.path.join(opts.linkjoe, os.path.basename(f))
-      if os.path.exists(joepath):
-        os.unlink(joepath)
-      os.link(f, joepath)
+      if not cuav_util.polygon_outside((pos.lat, pos.lon), boundary):
+        joepath = os.path.join(opts.linkjoe, os.path.basename(f))
+        if os.path.exists(joepath):
+          os.unlink(joepath)
+        os.symlink(f, joepath)
 
-    if opts.mosaic and len(regions) > 0:
-      composite = cuav_mosaic.CompositeThumbnail(img_scan, regions, quality=opts.quality)
-      chameleon.save_file('composite.jpg', composite)
-      thumbs = cuav_mosaic.ExtractThumbs(cv.LoadImage('composite.jpg'), len(regions))
+    if pos and len(regions) > 0:
       latlon_list = joelog.add_regions(frame_time, regions, pos, f)
       print pos, latlon_list
+
+    if opts.mosaic and len(regions) > 0:
+      composite = cuav_mosaic.CompositeThumbnail(im_full, regions, quality=opts.quality)
+      chameleon.save_file('composite.jpg', composite)
+      thumbs = cuav_mosaic.ExtractThumbs(cv.LoadImage('composite.jpg'), len(regions))
       mosaic.add_regions(regions, thumbs, latlon_list, f, pos)
     if pos:
       mosaic.add_image(f, img_scan, pos)
@@ -158,21 +195,20 @@ def process(files):
           x2 *= 2
           y2 *= 2
         cv.Rectangle(mat, (max(x1-2,0),max(y1-2,0)), (x2+2,y2+2), (255,0,0), 2)
-      cv.ShowImage('Viewer', mat)
-      cv.WaitKey(1)
-      cv.WaitKey(1)
+      cv.CvtColor(mat, mat, cv.CV_BGR2RGB)
+      viewer.set_image(mat)
 
     total_time += (t1-t0)
     print('%s scan %.1f fps  %u regions [%u/%u]' % (
       f, count/total_time, region_count, scan_count, num_files))
-    
 
-if opts.view:
-    cv.NamedWindow('Viewer')
 
 # main program
 state = state()
 
 process(args)
-cuav_util.cv_wait_quit()
-cv.DestroyWindow('Viewer')
+while True:
+  if opts.mosaic:
+    slipmap.check_events()
+  time.sleep(0.01)
+
